extracted redispatch test case into a separate test class to allow use of World rather than TestWorld
